Update bikeshare.py
docs: Update load_data and time_stats docstrings

Change the description of the month and day arguments in the docstrings for the load_data and time_stats functions.  This was to make clear that even though it will be passed in as a string, it is a string of an integer (or "all").