Update bikeshare.py
refactor: Renaming get_filters variables

The user_city_in, user_month_in and user_day_in variable names in the get_filters function were not inline with the naming conventions used in the rest of the code.  This was fixed by removing the "_in" from the end of the variable names.